+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Brave is now a single instance application.
 - HTTPS Everywhere information is now displayed in the site information popup.
 - Find in page now shows match ordinal.
+- Open all bookmarks in bookmark folder context menu added.
 - Upgraded to libchromiumcontent 49.0.2623.87.
 - Upgraded to Node v5.9.